add float integration

+++ internal/domain/route.go
@@ -31,7 +31,7 @@ type Route struct {
 	DBPlanDates     []DbPlanInfo    `json:"db_plan"`
 	BusyDates       []DbPlanInfo    `json:"busy_dates"`
 	Up              int             `json:"up" db:"up"`
-	Time            int             `json:"time" db:"time"`
+	Time            float64         `json:"time" db:"time"`
 	Adjustment      int             `json:"adjustment" db:"adjustment"`
 	Shift           string          `json:"shift" db:"shift"`
 	NeedShifts      int             `json:"need_shifts" db:"need_shifts"`
